+++ api/src/services/import.test.ts
@@ -2,7 +2,7 @@ import { PassThrough, Readable } from 'node:stream';
 import { setTimeout } from 'node:timers/promises';
 import { ErrorCode, ForbiddenError, isDirectusError } from '@directus/errors';
 import { SchemaBuilder } from '@directus/schema-builder';
-import type { Accountability, SchemaOverview } from '@directus/types';
+import type { Accountability, ImportCollectionData, SchemaOverview } from '@directus/types';
 import knex, { type Knex } from 'knex';
 import { createTracker, MockClient, Tracker } from 'knex-mock-client';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
@@ -10,7 +10,6 @@ import { getCache } from '../cache.js';
 import emitter from '../emitter.js';
 import { validateAccess } from '../permissions/modules/validate-access/validate-access.js';
 import { createDefaultAccountability } from '../permissions/utils/create-default-accountability.js';
-import type { ImportCollectionData } from '../utils/build-import-plan.js';
 import { getService } from '../utils/get-service.js';
 import { ImportService } from './import.js';
 import { NotificationsService } from './notifications.js';
