+++ packages/types/src/index.ts
@@ -16,7 +16,8 @@ export * from './folders.js';
 export * from './geometry.js';
 export * from './graphql.js';
 export * from './server.js';
-export * from './import-export.js';
+export * from './export.js';
+export * from './import.js';
 export * from './items.js';
 export * from './mail.js';
 export * from './misc.js';
